Fix: deleted old openAI test code and created actual ai code for finding gift suggestions with profile props being passed in.
